pyproject.toml
[tool.poetry]
name = "mp_helper"
version = "0.1.0"
description = ""
authors = ["hasbai <jsclndnz@outlook.com>"]
license = "Apache"

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.21.3"
fastapi = {extras = ["all"], version = "^0.73.0"}
diskcache = "^5.4.0"
Pillow = "^9.0.0"
orjson = "^3.6.6"
beautifulsoup4 = "^4.10.0"
lxml = "^4.7.1"
markdown2 = "^2.4.2"
selenium = "^4.1.0"
pyperclip = "^1.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
